Add temporal JSON support starting from the TJSONB fork

+++ meos/test/interval_test.c
@@ -65,10 +65,10 @@ int main(void)
 
   /* Create the result types for the bool functions of the API */
   bool bool_result;
-  int32 int32_result;
-  uint32 uint32_result;
-  uint64 uint64_result;
-  float8 float8_result;
+  int32_t int32_result;
+  uint32_t uint32_result;
+  uint64_t uint64_result;
+  double float8_result;
   char *char_result;
   Interval *interv_result;
   Numeric numeric_result;
@@ -85,13 +85,13 @@ int main(void)
   printf("add_interval_interval(%s, %s): %s\n", interv_out1, interv_out2, char_result);
   free(interv_result); free(char_result);
 
-  /* Interval *div_interval_float8(const Interval *interv, float8 factor); */
+  /* Interval *div_interval_float8(const Interval *interv, double factor); */
   interv_result = div_interval_float8(interv_in1, 5);
   char_result = interval_out(interv_result);
   printf("div_interval_float8(%s, 5): %s\n", interv_out1, char_result);
   free(interv_result); free(char_result);
 
-  /* int32 interval_cmp(const Interval *interv1, const Interval *interv2); */
+  /* int32_t interval_cmp(const Interval *interv1, const Interval *interv2); */
   int32_result = interval_cmp(interv_in1, interv_in2);
   printf("interval_cmp(%s, %s): %d\n", interv_out1, interv_out2, int32_result);
 
@@ -120,15 +120,15 @@ int main(void)
   bool_result = interval_gt(interv_in1, interv_in2);
   printf("interval_gt(%s, %s): %c\n", interv_out1, interv_out2, bool_result ? 't' : 'f');
 
-  /* uint32 interval_hash(const Interval *interv); */
+  /* uint32_t interval_hash(const Interval *interv); */
   uint32_result = interval_hash(interv_in1);
   printf("interval_hash(%s): %u\n", interv_out1, uint32_result);
 
-  /* uint64 interval_hash_extended(const Interval *interv, uint64 seed); */
+  /* uint64_t interval_hash_extended(const Interval *interv, uint64_t seed); */
   uint64_result = interval_hash_extended(interv_in1, 1);
   printf("interval_hash_extended(%s): %lu\n", interv_out1, uint64_result);
 
-  /* Interval *interval_in(const char *str, int32 typmod); */
+  /* Interval *interval_in(const char *str, int32_t typmod); */
   interv_result = interval_in("3 hours", -1);
   char_result = interval_out(interv_result);
   printf("interval_in(\"3 hours\", -1): %s\n", char_result);
@@ -170,7 +170,7 @@ int main(void)
   bool_result = interval_lt(interv_in1, interv_in2);
   printf("interval_lt(%s, %s): %c\n", interv_out1, interv_out2, bool_result ? 't' : 'f');
 
-  /* Interval *interval_make(int32 years, int32 months, int32 weeks, int32 days, int32 hours, int32 mins, float8 secs); */
+  /* Interval *interval_make(int32_t years, int32_t months, int32_t weeks, int32_t days, int32_t hours, int32_t mins, double secs); */
   interv_result = interval_make(2025, 3, 1, 4, 8, 0, 0);
   char_result = interval_out(interv_result);
   printf("interval_make(2025, 3, 1, 4, 8, 0, 0): %s\n", char_result);
@@ -191,13 +191,13 @@ int main(void)
   printf("interval_out(%s): %s\n", interv_out1, char_result);
   free(char_result);
 
-  /* float8 interval_part(const Interval *interv, const text *units); */
+  /* double interval_part(const Interval *interv, const text *units); */
   units = text_in("seconds");
   float8_result = interval_part(interv_in1, units);
   printf("interval_part(%s, \"seconds\"): %lf\n", interv_out1, float8_result);
   free(units);
 
-  /* Interval *interval_scale(const Interval *interv, int32 typmod); */
+  /* Interval *interval_scale(const Interval *interv, int32_t typmod); */
   /* The value below sets the precision to (0) */
   Interval *interv = interval_in("1 hour 1 minute 1.123456789 seconds", 2147418112);
   char_result = interval_out(interv);
@@ -223,13 +223,13 @@ int main(void)
   printf("minus_interval_interval(%s, %s): %s\n", interv_out1, interv_out2, char_result);
   free(interv_result); free(char_result);
 
-  /* Interval *mul_float8_interval(float8 factor, const Interval *interv); */
+  /* Interval *mul_float8_interval(double factor, const Interval *interv); */
   interv_result = mul_float8_interval(8, interv_in1);
   char_result = interval_out(interv_result);
   printf("mul_float8_interval(8, %s): %s\n", interv_out1, char_result);
   free(interv_result); free(char_result);
 
-  /* Interval *mul_interval_float8(const Interval *interv, float8 factor); */
+  /* Interval *mul_interval_float8(const Interval *interv, double factor); */
   interv_result = mul_interval_float8(interv_in1, 8);
   char_result = interval_out(interv_result);
   printf("mul_interval_float8(%s, 8): %s\n", interv_out1, char_result);
